requirements.txt
openpyxl==3.1.5
pydantic>=2.7,<3
PyYAML>=6.0,<7
pandas>=2.2,<3
numpy>=2.0,<3
jsonschema>=4.25,<5
xlwings>=0.33,<1
mcp>=1.0.0
